+++ android/src/main/java/com/splunkotelreactnative/crash/CrashReporterAttributes.java
@@ -8,7 +8,7 @@ import io.opentelemetry.api.common.AttributeKey;
 
 // copied from splunk-otel-android SplunkRum class
 class CrashReporterAttributes {
-  static final AttributeKey<String> SPLUNK_OPERATION_KEY = stringKey("_splunk_operation");
+  static final AttributeKey<String> SPLUNK_OPERATION_KEY = stringKey("_hyperdx_operation");
   static final AttributeKey<Long> STORAGE_SPACE_FREE_KEY = longKey("storage.free");
   static final AttributeKey<Long> HEAP_FREE_KEY = longKey("heap.free");
   static final AttributeKey<Double> BATTERY_PERCENT_KEY = doubleKey("battery.percent");