+++ example/src/App.tsx
@@ -26,16 +26,16 @@ import {
   OtelWrapper,
   SplunkRum,
   startNavigationTracking,
-} from '@splunk/otel-react-native';
-import type { ReactNativeConfiguration } from '@splunk/otel-react-native';
+} from '@hyperdx/otel-react-native';
+import type { ReactNativeConfiguration } from '@hyperdx/otel-react-native';
 import Config from 'react-native-config';
 
 const RumConfig: ReactNativeConfiguration = {
   //TODO fix config setting for iOS in inegration tests
-  beaconEndpoint: Config.BEACON_ENDPOINT || 'http://localhost:53820/zipkindump',
-  applicationName: 'RnExample',
   allowInsecureBeacon: true,
-  rumAccessToken: '',
+  beaconEndpoint: Config.BEACON_ENDPOINT || 'http://localhost:53820/zipkindump',
+  service: 'RnExample',
+  apiKey: 'superkey-1234',
   debug: true,
   globalAttributes: {
     'app.version': '1.1.5',