+++ CHANGELOG.md
@@ -1,6 +1,31 @@
-# Changelog
+# @hyperdx/otel-react-native
 
-## Unreleased
+## 0.2.0-next.0
+
+### Minor Changes
+
+- chore: merge latest upstream
+
+## 0.3.4
+
+* Add Privacyinfo for iOS
+
+## 0.3.3
+
+* Fix session timeout causing a crash
+
+## 0.3.2
+
+* Upgrade OTEL components to 1.20
+* Fix iOS number conversion
+
+## 0.3.1
+
+* Fix global attributes
+
+## 0.3.0
+
+* Add carrier info
 
 ## 0.3.4
 
@@ -25,4 +50,10 @@
 
 ## 0.2.0
 
-* Add disk caching for java
+- Add disk caching for java
+
+## 0.1.0
+
+### Minor Changes
+
+- chore: release v0.1.0